Drop Python 3.4 support; add support for Python 3.7 and above
