+++ tox.ini
@@ -1,0 +1,7 @@
+[tox]
+envlist = py35,py36,py37,py38
+
+[testenv]
+deps = pytest
+commands =
+    pytest
